Refactor API for reusing resource operations
