+++ pkg/api/api.go
@@ -4,50 +4,17 @@ package api
 
 import (
 	"context"
-	"fmt"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
+	"github.com/grafana/xk6-kubernetes/pkg/resources"
 
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/runtime/serializer/yaml"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/rest"
 )
 
-// maps kinds to api resources
-// TODO: complete with most common kinds
-func knownKinds(kind string) (schema.GroupVersionResource, error) {
-	kindMapping := map[string]schema.GroupVersionResource{
-		"ConfigMap":             {Group: "", Version: "v1", Resource: "configmaps"},
-		"Deployment":            {Group: "apps", Version: "v1", Resource: "deployments"},
-		"Job":                   {Group: "batch", Version: "v1", Resource: "jobs"},
-		"PersistentVolume":      {Group: "", Version: "v1", Resource: "persistentvolumes"},
-		"PersistentVolumeClaim": {Group: "", Version: "v1", Resource: "persistentvolumeclaims"},
-		"Pod":                   {Group: "", Version: "v1", Resource: "pods"},
-		"Namespace":             {Group: "", Version: "v1", Resource: "namespaces"},
-		"Node":                  {Group: "", Version: "v1", Resource: "nodes"},
-		"Secret":                {Group: "", Version: "v1", Resource: "secrets"},
-		"Service":               {Group: "", Version: "v1", Resource: "services"},
-		"StatefulSet":           {Group: "apps", Version: "v1", Resource: "statefulsets"},
-	}
-
-	gvk, found := kindMapping[kind]
-	if !found {
-		return schema.GroupVersionResource{}, fmt.Errorf("unknown kind: '%s'", kind)
-	}
-	return gvk, nil
-}
-
 // Kubernetes defines an interface that extends kubernetes interface[k8s.io/client-go/kubernetes.Interface] adding
 // generic functions that operate on any kind of object
 type Kubernetes interface {
-	Apply(manifest string) error
-	Create(obj map[string]interface{}) (map[string]interface{}, error)
-	Get(kind string, name string, namespace string) (map[string]interface{}, error)
-	List(kind string, namespace string) ([]map[string]interface{}, error)
-	Delete(kind string, name string, namespace string) error
+	resources.Operations
 }
 
 // KubernetesConfig defines the configuration for creating a Kubernetes instance
@@ -60,145 +27,32 @@ type KubernetesConfig struct {
 	Client dynamic.Interface
 }
 
-// kubernetes holds the reference to the helpers for interacting with kubernetes
+// kubernetes holds references to implementation of the Kubernetes interface
 type kubernetes struct {
-	ctx        context.Context
-	client     dynamic.Interface
-	serializer runtime.Serializer
+	ctx context.Context
+	*resources.Client
 }
 
 // NewFromConfig returns a Kubernetes instance
 func NewFromConfig(c KubernetesConfig) (Kubernetes, error) {
-	client := c.Client
-	var err error
-	if client == nil {
-		client, err = dynamic.NewForConfig(c.Config)
-		if err != nil {
-			return nil, err
-		}
-	}
-
 	ctx := c.Context
 	if ctx == nil {
 		ctx = context.TODO()
 	}
 
+	var client *resources.Client
+	var err error
+	if c.Client != nil {
+		client = resources.NewFromClient(ctx, c.Client)
+	} else {
+		client, err = resources.NewFromConfig(ctx, c.Config)
+		if err != nil {
+			return nil, err
+		}
+	}
+
 	return &kubernetes{
-		ctx:        ctx,
-		client:     client,
-		serializer: yaml.NewDecodingSerializer(unstructured.UnstructuredJSONScheme),
+		ctx:    ctx,
+		Client: client,
 	}, nil
-}
-
-// Apply creates a resource in a kubernetes cluster from a YAML manifest
-func (k *kubernetes) Apply(manifest string) error {
-	uObj := &unstructured.Unstructured{}
-	_, gvk, err := k.serializer.Decode([]byte(manifest), nil, uObj)
-	if err != nil {
-		return err
-	}
-	resource, err := knownKinds(gvk.Kind)
-	if err != nil {
-		return err
-	}
-
-	namespace := uObj.GetNamespace()
-	if namespace == "" {
-		namespace = "default"
-	}
-
-	_, err = k.client.Resource(resource).
-		Namespace(namespace).
-		Create(
-			k.ctx,
-			uObj,
-			metav1.CreateOptions{},
-		)
-	return err
-}
-
-// Create creates a resource in a kubernetes cluster from an object with its specification
-func (k *kubernetes) Create(obj map[string]interface{}) (map[string]interface{}, error) {
-	uObj := &unstructured.Unstructured{
-		Object: obj,
-	}
-
-	gvk := uObj.GroupVersionKind()
-	namespace := uObj.GetNamespace()
-	if namespace == "" {
-		namespace = "default"
-	}
-	resource, err := knownKinds(gvk.Kind)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := k.client.Resource(resource).
-		Namespace(namespace).
-		Create(
-			k.ctx,
-			uObj,
-			metav1.CreateOptions{},
-		)
-	if err != nil {
-		return nil, err
-	}
-	return resp.UnstructuredContent(), nil
-}
-
-// Get returns an object given its kind, name and namespace
-func (k *kubernetes) Get(kind string, name string, namespace string) (map[string]interface{}, error) {
-	resource, err := knownKinds(kind)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := k.client.
-		Resource(resource).
-		Namespace(namespace).
-		Get(
-			k.ctx,
-			name,
-			metav1.GetOptions{},
-		)
-	if err != nil {
-		return nil, err
-	}
-	return resp.UnstructuredContent(), nil
-}
-
-// List returns a list of objects given its kind and namespace
-func (k *kubernetes) List(kind string, namespace string) ([]map[string]interface{}, error) {
-	resource, err := knownKinds(kind)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := k.client.
-		Resource(resource).
-		Namespace(namespace).
-		List(k.ctx, metav1.ListOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	list := []map[string]interface{}{}
-	for _, uObj := range resp.Items {
-		list = append(list, uObj.UnstructuredContent())
-	}
-	return list, nil
-}
-
-// Delete deletes an object given its kind, name and namespace
-func (k *kubernetes) Delete(kind string, name string, namespace string) error {
-	resource, err := knownKinds(kind)
-	if err != nil {
-		return err
-	}
-
-	err = k.client.
-		Resource(resource).
-		Namespace(namespace).
-		Delete(k.ctx, name, metav1.DeleteOptions{})
-
-	return err
 }
